Fix display of colors when wrapping text
When colored text wraps to the next line, the color was only applied
until the end of the line, the text continued on the following line was
reset to default color.

This reapplies the currently used color at the beginning of each wrapped
line, unless it has been properly reset.